backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
bcrypt==4.0.1
pydantic==2.2.1
Werkzeug==2.3.7
pytest==7.4.0
pytest-mock==3.11.1
pytest-cov==4.1.0
